third_party/WebKit/LayoutTests/compositing/dont-composite-select-elements.html

<!doctype HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<select size="2">
  <option> value 1</option>
  <option> value 2</option>
  <option> value 3</option>
  <option> value 4</option>
</select>
<script>
if (window.internals)
  window.internals.settings.setPreferCompositingToLCDTextEnabled(true);
if (window.testRunner) {
  testRunner.dumpAsText();
}
test(function() {
    var json = JSON.parse(window.internals.layerTreeAsText(document));
    // The <select> element's scroller should be painted into the root layer.
    assert_true(json["layers"].length == 1);
    }, "test");
</script>
